Fetch events through redux. Time selection isn't implemented yet
