+++ package.json
@@ -44,6 +44,7 @@
     "react-tap-event-plugin": "^0.2.1",
     "redux": "^3.0.5",
     "redux-simple-router": "^1.0.2",
+    "redux-thunk": "^1.0.2",
     "typeahead.js": "^0.11.1",
     "underscore": "^1.8.3"
   },
