+++ src/actions/events.js
@@ -1,10 +1,23 @@
-const constants = require('../constants');
+import constants from '../constants'
+import fetch from 'isomorphic-fetch'
 
-function getEvents(q) {
-  return {
-    type: constants.GET_EVENTS,
-    query: q
-  };
+function makeRequest(query) {
+    var url = `${appSettings.api_base}/event/?text=${query}`
+    return fetch(url)
 }
 
-module.exports = { increase, decrease };
+export function receiveEvents(json) {
+    return {
+        type: constants.RECEIVE_EVENTS,
+        items: json.data,
+        receivedAt: Date.now()
+    }
+}
+
+export function fetchEvents(query) {
+    return (dispatch) => {
+        return makeRequest(query)
+            .then(response => response.json())
+            .then(json => dispatch(receiveEvents(json)))
+    }
+}
